sizeof op

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int n;
 
 	/*the sizeof operator returns the sizeof a data type*/
-	printf("size of char: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a float type: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(n));
+	printf("size of char: %lu byte(s)\n", sizeof(d));
+	printf("size of an int: %lu byte(s)\n", sizeof(b));
+	printf("size of a long int: %lu byte(s)\n", sizeof(c));
+	printf("size of a float type: %lu byte(s)\n", sizeof(d));
+	printf("size of a long long int: %lu byte(s)\n", sizeof(n));
 
 	return (0);
 }
